feat : blockButton을 fakeBox 내부로 옮겨 이벤트가 잘 작동하도록 변경
